refactor(model): switch to using database for data management
